requirements.txt
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
coverage==5.3
cycler==0.10.0
decorator==4.4.2
et-xmlfile==1.1.0
ExifRead==2.3.2
fiona==1.9.5
fonttools==4.28.5
future==0.18.2
GDAL==3.8.4
geographiclib==1.50
geojson==3.1.0
geopandas==0.14.4
httplib2==0.18.1
idna==2.10
importlib-metadata==7.0.1
Jinja2==3.1.2
kiwisolver==1.2.0
lxml==4.6.2
MarkupSafe==2.1.5
matplotlib==3.5.1
mock==4.0.2
networkx==2.5
nose2==0.9.2
numpy==1.24.1
openpyxl==3.0.9
OWSLib==0.29.2
packaging==23.2
pandas==2.0.2
Pillow==10.0.1
plotly==4.12.0
ply==3.11
psycopg2-binary==2.9.3
Pygments==2.17.2
pyodbc==4.0.30
PyOpenGL==3.1.5
pyparsing==2.4.7
PyPDF2==3.0.1
pypiwin32==223
pyproj==3.6.1
PyQt5==5.15.9
PyQt5-sip==12.13.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-ulid==2.7.0
pytz==2023.3
pywin32==306
PyYAML==5.3.1
remotior_sensus==0.3.5
reportlab==4.0.4
requests==2.31.0
retrying==1.3.3
Rtree==1.2.0
scipy==1.10.1
setuptools==67.6.0
shapely==2.0.2
simplejson==3.17.2
sip==6.1.1
six==1.15.0
toml==0.10.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.25.11
wxPython==4.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0














